+++ lib/templates/new/app/routes/healthz.rb
@@ -1,0 +1,15 @@
+module Routes
+  class Healthz < App
+
+    route do |r|
+
+      # GET /healthz
+      r.root do
+        response['Content-Type'] = 'text/plain; charset=utf-8'
+        'alive and kicking'
+      end
+
+    end
+
+  end
+end